Integrate rapidjson into project
