+++ include/cppinclude/linkedList_rand.cc
@@ -1,0 +1,20 @@
+#include "linkedList_rand.h"
+#include "rapidjson/document.h"
+
+// transform the json representation of linked list into the linked list
+Node*
+LinkedListRandom::list2list(const char*  json) {
+  
+}
+
+// transform a linked list to the json representation of linked list
+std::string
+LinkedListRandom::printList(Node* head) {
+  
+}
+
+// delete the linked list
+void
+LinkedListRandom::freeList(Node* &head) {
+  
+}
